feat: Styled report page (#125)
