--- frontend/pages/panel/report/[id].vue
@@ -1,113 +1,0 @@
-<template>
-  <div>
-    <h1>Report</h1>
-    <Map :tile-layer-overlay="selectedMosaic" />
-    <Button @click="selectMosaic(0)">
-      Natural Color
-    </Button>
-    <Button @click="selectMosaic(1)">
-      Temperature
-    </Button>
-    <Button @click="selectMosaic(3)">
-      Agriculture
-    </Button>
-    <Button @click="selectMosaic(4)">
-      Vegetation
-    </Button>
-    <Button @click="selectMosaic(5)">
-      Moisture
-    </Button>
-    <Button @click="selectMosaic(6)">
-      Atmospheric Penetration
-    </Button>
-    <Table>
-      <TableHeader>
-        <TableRow>
-          <TableHead class="w-[150px]">
-            Data
-          </TableHead>
-          <TableHead>Value</TableHead>
-        </TableRow>
-      </TableHeader>
-      <TableBody>
-        <TableRow v-for="(value, key) in metadata" :key="key">
-          <TableCell class="font-medium">
-            {{ key }}
-          </TableCell>
-          <TableCell>
-            {{ value }}
-          </TableCell>
-        </TableRow>
-      </TableBody>
-    </Table>
-    <LineChart
-      :data="reportData.temperature_chart"
-      index="temperature"
-      :categories="['distribution']"
-      :colors="['#13b982']"
-    />
-    <LineChart
-      :data="reportData.reflectance_chart"
-      index="wave_length"
-      :categories="['reflectance']"
-      :colors="['#13b982']"
-    />
-  </div>
-</template>
-
-<script setup lang="ts">
-import type { LatLngBoundsLiteral } from "leaflet";
-import { LineChart } from "~/components/ui/chart-line";
-
-const route = useRoute();
-const reportData: any = ref(null);
-const selectedMosaic = ref<{
-  url: string
-  bounds: LatLngBoundsLiteral
-} | undefined>(undefined);
-
-async function selectMosaic(type: number) {
-  const { data } = await useApi("/landsat/mosaic", {
-    query: {
-      scene_id: reportData.value.id,
-      collection_id: "landsat-c2-l2",
-      mosaic_type: type.toString() as any
-    }
-  });
-  let url: string = data.value!.toString();
-  url = url.replace("%7Bz%7D", "{z}");
-  url = url.replace("%7Bx%7D", "{x}");
-  url = url.replace("%7By%7D", "{y}");
-  console.log(url);
-  selectedMosaic.value = {
-    url,
-    bounds: reportData.value.polygon.coordinates.map(({ latitude, longitude }) => [latitude, longitude] as [number, number])
-  };
-}
-
-const { data, error } = await useApi("/report/get_report", {
-  query: {
-    scene_id: route.params.id.toString()
-  }
-});
-
-if (error.value) {
-  console.error("Error fetching report:", error.value);
-} else if (data.value) {
-  reportData.value = data.value;
-}
-
-const metadata = {
-  "Scene ID": reportData.value.id,
-  Satellite: reportData.value.platform.replace("-", " ").replace("landsat", "Landsat"),
-  "Cloud Coverage": reportData.value.eo_cloud_cover,
-  Path: reportData.value.wrs_coordinates.wrs_path,
-  Row: reportData.value.wrs_coordinates.wrs_row,
-  "Collection Category": reportData.value.metadata.collection_category,
-  "Collection Number": reportData.value.metadata.collection_number,
-  "Created At": reportData.value.metadata.created,
-  "Sun Azimuth": reportData.value.metadata.sun_azimuth,
-  "Sun Elevation": reportData.value.metadata.sun_elevation,
-  Instruments: reportData.value.metadata.instruments.join(", ")
-};
-</script>
